+++ js/data.js
@@ -1,0 +1,25 @@
+const DESCRIPTIONS = [
+  'Наш санаторий в Сочи #отпуск #отдыхаем',
+  'Путь на пляж.. #beach #отдых #море',
+  'Незабываемый отдых #maledives #мечта',
+  'Устроили фотосессию #фоткаемся #лето',
+  'Веселенький обед)))',
+  'МОТИВАЦИЯ #машинамечты',
+  'Снова диета к лету... #пп #худею',
+  'Нужно радовать себя каждый день и исполнять свои мечты <3 #сбывшаясямечта',
+  'Купили модные обновочки',
+  'Отдохнули с друзьями)))))'
+];
+
+const NAMES = ['Иван', 'Мария', 'Виктор', 'Юлия', 'Игорь', 'Елена'];
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {DESCRIPTIONS, NAMES, COMMENT_MESSAGES};
